Move get_hess and clean_hess to hesstools and make the necessary changes so they can be called by other parts of the code.
